Import Actions too in fixture importer
