+++ tools/fixture_importer/importer.go
@@ -10,21 +10,14 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
-	"net/url"
-	"strconv"
-	"strings"
-	"time"
-
-	spew "github.com/davecgh/go-spew/spew"
-
 	apiclient "github.com/creativesoftwarefdn/weaviate/client"
-	"github.com/creativesoftwarefdn/weaviate/client/things"
 	"github.com/creativesoftwarefdn/weaviate/models"
-
 	"github.com/go-openapi/runtime"
 	httptransport "github.com/go-openapi/runtime/client"
 	"github.com/go-openapi/strfmt"
+	"io/ioutil"
+	"net/url"
+	"strings"
 )
 
 var APITOKEN string
@@ -40,8 +33,8 @@ var schema Schema
 var weaviateUrl url.URL
 
 type DemoDataset struct {
-	Things []DemoInstance `json:"Things"`
-	//Actions []DemoInstance `json:"Actions"`
+	Things  []DemoInstance `json:"Things"`
+	Actions []DemoInstance `json:"Actions"`
 }
 
 type DemoInstance map[string]interface{}
@@ -51,144 +44,31 @@ var auth runtime.ClientAuthInfoWriterFunc
 
 var demoDataset DemoDataset
 
+type fixupAddRef struct {
+	kind         string
+	fromId       string
+	fromProperty string
+	toClass      string
+	toId         string
+}
+
+var idMap map[string]string
+var thingFixups []fixupAddRef
+var actionFixups []fixupAddRef
+
+var classKinds map[string]string
+
 func main() {
+	idMap = map[string]string{}
+	thingFixups = []fixupAddRef{}
+	actionFixups = []fixupAddRef{}
+	classKinds = map[string]string{}
+
 	createThings()
-}
+	createActions()
 
-func createThings() {
-	type fixupAddRef struct {
-		fromId       string
-		fromProperty string
-		toClass      string
-		toId         string
-	}
-
-	// need to fix up data later.
-	fixups := []fixupAddRef{}
-
-	// Map ID's to UUID's
-	idMap := map[string]string{}
-
-	for _, thing := range demoDataset.Things {
-		className := thing["class"].(string)
-		uuid := thing["uuid"].(string)
-
-		properties := map[string]interface{}{}
-
-		for key, value := range thing {
-			if key == "class" || key == "uuid" {
-				continue
-			}
-
-			ref, isRef := value.(map[string]interface{})
-			if isRef {
-				fixups = append(fixups, fixupAddRef{
-					fromId:       uuid,
-					fromProperty: key,
-					toClass:      ref["class"].(string),
-					toId:         ref["uuid"].(string),
-				})
-			} else {
-				class := findClass(schema.Things, className)
-				property := findProperty(class, key)
-				if len(property.AtDataType) != 1 {
-					panic(fmt.Sprintf("Only one datatype supported for import. Failed in thing %s.%s", className, property.Name))
-				}
-				dataType := property.AtDataType[0]
-
-				switch dataType {
-				case "string":
-					properties[key] = value
-				case "int":
-					v, err := strconv.ParseInt(value.(string), 10, 64)
-					if err != nil {
-						panic(err)
-					}
-					properties[key] = v
-				case "number":
-					properties[key] = value.(float64)
-				case "boolean":
-					properties[key] = value.(bool)
-				default:
-					panic(fmt.Sprintf("No such datatype supported: %s", dataType))
-				}
-			}
-		}
-
-		t := models.ThingCreate{
-			AtContext: "http://example.org",
-			AtClass:   className,
-			Schema:    properties,
-		}
-
-		thing := assertCreateThing(&t)
-		idMap[uuid] = string(thing.ThingID) // Store mapping of ID's
-		fmt.Printf("Created Thing %s\n", thing.ThingID)
-	}
-
-	fmt.Printf("Checking if all things that need a patch are created.\n")
-	for {
-		allExist := true
-
-		for _, fixup := range fixups {
-			if !checkThingExists(idMap[fixup.fromId]) {
-				allExist = false
-				fmt.Printf("From does not exist! %v\n", idMap[fixup.fromId])
-			}
-			if !checkThingExists(idMap[fixup.toId]) {
-				allExist = false
-				fmt.Printf("To does not exist! %v\n", idMap[fixup.toId])
-				break
-			}
-		}
-
-		if allExist {
-			fmt.Printf("Everything that needs to be patched exists!\n")
-			break
-		} else {
-			fmt.Printf("Not everything that needs to be patched exists\n")
-
-			var waitSecondsUntilSettled time.Duration = 2 * time.Second
-			fmt.Printf("Waiting for %v to settle\n", waitSecondsUntilSettled)
-			time.Sleep(waitSecondsUntilSettled)
-			continue
-		}
-	}
-
-	// Now fix up refs
-	op := "add"
-	for _, fixup := range fixups {
-		path := fmt.Sprintf("/schema/%s", fixup.fromProperty)
-		patch := &models.PatchDocument{
-			Op:   &op,
-			Path: &path,
-			Value: map[string]interface{}{
-				"$cref":       idMap[fixup.toId],
-				"locationUrl": "http://localhost:8080",
-				//"type":        fixup.toClass,
-				"type": "Thing",
-			},
-		}
-
-		assertPatchThing(idMap[fixup.fromId], patch)
-		fmt.Printf("Patched %s\n", idMap[fixup.fromId])
-	}
-}
-
-func checkThingExists(id string) bool {
-	params := things.NewWeaviateThingsGetParams().WithThingID(strfmt.UUID(id))
-	resp, err := client.Things.WeaviateThingsGet(params, auth)
-
-	if err != nil {
-		switch v := err.(type) {
-		case *things.WeaviateThingsGetNotFound:
-			return false
-		default:
-			panic(fmt.Sprintf("Can't create thing %#v, because %#v", resp, spew.Sdump(v)))
-		}
-	}
-
-	return true
+	fixupThings()
+	fixupActions()
 }
 
 func findClass(schema *models.SemanticSchema, className string) *models.SemanticSchemaClass {
@@ -218,43 +98,6 @@ func joinErrorMessages(errResponse *models.ErrorResponse) string {
 	}
 
 	return strings.Join(msgs, ", ")
-}
-
-func assertCreateThing(t *models.ThingCreate) *models.ThingGetResponse {
-	params := things.NewWeaviateThingsCreateParams().WithBody(things.WeaviateThingsCreateBody{Thing: t})
-
-	resp, _, err := client.Things.WeaviateThingsCreate(params, auth)
-
-	if err != nil {
-		switch v := err.(type) {
-		case *things.WeaviateThingsCreateUnprocessableEntity:
-			panic(fmt.Sprintf("Can't create thing %#v, because %s", t, joinErrorMessages(v.Payload)))
-		default:
-			panic(fmt.Sprintf("Can't create thing %#v, because %#v", t, spew.Sdump(err)))
-		}
-	}
-
-	return resp.Payload
-}
-
-func assertPatchThing(id string, p *models.PatchDocument) *models.ThingGetResponse {
-	params := things.NewWeaviateThingsPatchParams().WithBody([]*models.PatchDocument{p}).WithThingID(strfmt.UUID(id))
-
-	resp, err := client.Things.WeaviateThingsPatch(params, auth)
-
-	if err != nil {
-		switch v := err.(type) {
-		case *things.WeaviateThingsPatchNotFound:
-			panic(fmt.Sprintf("Can't patch thing with %s, because thing cannot be found", spew.Sdump(p)))
-		case *things.WeaviateThingsPatchUnprocessableEntity:
-			panic(fmt.Sprintf("Can't patch thing, because %s", joinErrorMessages(v.Payload)))
-		default:
-			_ = v
-			panic(fmt.Sprintf("Can't patch thing with %#v, because %#v", p, spew.Sdump(err)))
-		}
-	}
-
-	return resp.Payload
 }
 
 func loadSemantisSchema(path string) *models.SemanticSchema {
